# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MDDB API
  description: |
    MDDB (Markdown Database) is a high-performance, version-controlled markdown database with triple protocol support.

    ## Features
    - Store and manage markdown documents with metadata
    - Full revision history for all documents
    - Multi-language support
    - Fast metadata-based search
    - **Vector search** with embedding providers (OpenAI, Cohere, Voyage AI, Ollama)
    - **Full-text search** with TF-IDF, BM25, BM25F, PMISparse scoring, 7 search modes (simple, boolean, phrase, wildcard, proximity, range, fuzzy), and multi-language stemming (18 languages)
    - **Hybrid search** combining FTS + vector search with alpha blending or RRF fusion
    - **Webhooks** for real-time event notifications
    - **TTL** (time-to-live) support for automatic document expiration
    - **JSON Schema validation** for document metadata
    - **URL import** with YAML frontmatter parsing
    - **Authentication** - JWT tokens and API keys
    - **Authorization** - Collection-level RBAC (Read/Write/Admin)
    - **Memory RAG** - Conversational memory system with session management, message storage, semantic/hybrid recall, and summarization
    - **User Management** - Multi-user with admin roles
    - **Group Permissions** - Organize users into groups
    - Template variable substitution
    - Export capabilities (NDJSON, ZIP)
    - Backup and restore operations
...

    ## Protocols
    - **HTTP/JSON** - Port 11023 (documented here)
...
servers:
...
paths:
...
  /v1/add:
    post:
      tags:
        - Documents
      summary: Add or update document
      description: |
        Add a new document or update an existing one. If a document with the same collection, key, and language exists, it will be updated and a new revision will be created.
      operationId: addDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRequest"
      responses:
        "200":
          description: Document added/updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/delete-collection:
    post:
      tags:
        - Documents
      summary: Delete collection
      description: Delete all documents in a collection
      operationId: deleteCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCollectionRequest"
      responses:
        "200":
          description: Collection deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    description: Number of documents deleted
                    example: 42
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          description: Auto-generated document ID
          example: doc|blog|hello-world|en_US
        key:
          type: string
          description: Document key (unique within collection and language)
          example: hello-world
        lang:
          type: string
          description: Language code (e.g., en_US, pl_PL, de_DE)
          example: en_US
        meta:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Metadata key-value pairs (values are arrays)
          example:
            category: [ blog, tutorial ]
            author: [ John Doe ]
            tags: [ markdown, database ]
        contentMd:
          type: string
          description: Markdown content
          example: "# Hello World\n\nWelcome to MDDB!"
        addedAt:
          type: integer
          format: int64
          description: Unix timestamp when document was created
          example: 1704844800
        updatedAt:
          type: integer
          format: int64
          description: Unix timestamp when document was last updated
          example: 1704931200
    AddRequest:
      type: object
      required:
        - collection
        - key
        - lang
        - contentMd
      properties:
        collection:
          type: string
          description: Collection name
          example: blog
        key:
          type: string
          description: Document key
          example: hello-world
        lang:
          type: string
          description: Language code
          example: en_US
        meta:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Metadata key-value pairs
          example:
            category: [ blog ]
            author: [ John Doe ]
        contentMd:
          type: string
          description: Markdown content
          example: "# Hello World\n\nWelcome to MDDB!"
...
    DeleteCollectionRequest:
      type: object
      required:
        - collection
      properties:
        collection:
          type: string
          description: Collection name to delete
          example: blog
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: document not found
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /v1/auth/login
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from POST /v1/auth/api-key